Refactor useAsync hook and add types
